Separate IPC functions
