+++ app/index.html
@@ -222,6 +222,7 @@
 <script src="js/sunburst3d.js"></script>
 <script src="js/treemap.js"></script>
 <script src="js/duFromFile.js"></script>
+<script src="js/ipc.js"></script>
 <script src="js/radar.js"></script>
 <script src="js/menu.js"></script>
 <script src="js/breadcrumbs.js"></script>
