+++ app/js/radar.js
@@ -1,12 +1,13 @@
 'use strict'
 
-const { remote, shell } = require('electron')
+const { shell } = require('electron')
 const path = require('path')
 
 const LASTLOAD_FILE = path.join(__dirname, 'lastload.json')
-// const child_process = require('child_process')
 
 // IPC handling
+// const { sendIpcMsg } = require(path.join(__dirname, 'js/ipc'))
+
 var current_size = 0,
   start_time
 
@@ -123,50 +124,6 @@ function complete(json) {
   // shell.beep() // disabling as this can be anonying for memory monitor
 }
 
-const DEBUG = process.env.DEBUG //
-const ipc_name = 'viz'
-const fs = require('fs')
-const zlib = require('zlib')
-
-let win
-
-var main_ipc = remote.ipcMain
-
-main_ipc.on('call', function(event, cmd) {
-  var args = Array.prototype.slice.call(arguments, 2)
-  handleIPC(cmd, args)
-})
-
-function setupRemoteIPC() {
-  win = new remote.BrowserWindow(DEBUG ? { width: 800, height: 600 } : { show: false })
-  win.loadURL('file://' + __dirname + '/headless.html')
-  if (DEBUG) win.openDevTools()
-
-  win.webContents.on('did-finish-load', function() {
-    // win.webContents.send('ready')
-    ready()
-  })
-}
-
-function setupWebViewIPC() {
-  webview.addEventListener('dom-ready', function() {
-    if (DEBUG) {
-      webview.openDevTools()
-    }
-    webview.style.display = 'none'
-  })
-
-  webview.addEventListener('ipc-message', function(event) {
-    var args = event.args
-    // var cmd = event.channel
-    var cmd = args.shift()
-    handleIPC(cmd, args)
-  })
-
-  // this triggers ready
-  webview.addEventListener('did-finish-load', ready)
-}
-
 function handleIPC(cmd, args) {
   switch (cmd) {
     case 'progress':
@@ -203,91 +160,6 @@ function fsipc(filename) {
     }
     console.timeEnd('fsipc')
   })
-}
-
-function setupIPC() {
-  var ipc = require('ipc')
-
-  ipc.send('register', ipc_name)
-
-  ipc.on('message', function(arg) {
-    log('[' + ipc_name + '] message', arg)
-  })
-
-  ipc.on('ready', function(arg) {
-    log('lets go')
-  })
-
-  // ipc listeners
-  ipc.on('progress', progress)
-  ipc.on('refresh', refresh)
-  ipc.on('complete', complete)
-}
-
-function setupLocalStorageIPC() {
-  window.addEventListener('storage', function(e) {
-    if (e.key == 'lsipc') {
-      var args = JSON.parse(e.newValue)
-      var cmd = args.shift()
-      handleIPC(cmd, args)
-    }
-  })
-}
-
-window.onbeforeunload = function(e) {
-  console.log('Closing time!!!!')
-  if (win) win.close()
-  // the better method would be to track client from
-  // browser.on('closed')
-}
-
-var child
-function setupChildIPC() {
-  console.log(path.join(__dirname, 'js/scanner.js'))
-  log('process.execArgv', process.execArgv, 'execPath', process.execPath)
-  child = child_process.fork('./js/scanner.js', {
-    env: process.ENV,
-    silent: true
-  })
-
-  // child = child_process.spawn(process.execPath, ['./js/scanner.js'], {
-  //   env: process.ENV,
-  //   stdio: ['pipe', 'pipe', 'pipe', 'ipc']
-  // })
-
-  child.on('message', function(args) {
-    var cmd = args.shift()
-    handleIPC(cmd, args)
-  })
-
-  // child.on('error', console.log.bind(console))
-
-  child.stdout.on('data', function(data) {
-    console.log('stdout: ' + data)
-  })
-
-  child.stderr.setEncoding('utf8')
-
-  child.stderr.on('data', function(data) {
-    console.log('stdout: ' + data)
-  })
-
-  child.on('close', function(code) {
-    console.log('child process exited with code ' + code)
-  })
-}
-
-setupLocalStorageIPC()
-// setupWebViewIPC()
-// setupChildIPC()
-setupRemoteIPC()
-// setupIPC()
-// ready() // run this
-
-function sendIpcMsg(cmd, msg) {
-  // webview.send('scan', msg)
-  // child.send({cmd: cmd, msg: msg})
-  win.webContents.send('scan', msg)
 }
 
 function ready() {
